refactor: extract subway api logic
