+++ src/js/subway/utils/api/stationManage.js
@@ -1,0 +1,71 @@
+import { request } from '../../../@shared/utils';
+import { BASE_URL } from '../../constants/constants';
+
+const getStations = async accessToken => {
+  if (!accessToken) return [];
+  const url = `${BASE_URL}/stations`;
+  const option = {
+    method: 'GET',
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      'Content-Type': 'application/json',
+    },
+  };
+
+  try {
+    const response = await request(url, option);
+    const stations = await response.json();
+
+    return stations;
+  } catch (error) {
+    throw new Error(error);
+  }
+};
+
+const addStation = async (accessToken, $input) => {
+  const url = `${BASE_URL}/stations`;
+  const option = {
+    method: 'POST',
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({
+      name: $input.value,
+    }),
+  };
+
+  try {
+    const response = await request(url, option);
+
+    return await response.json();
+  } catch (error) {
+    throw new Error(error.message);
+  }
+};
+
+const modifyStation = async () => {};
+
+const removeStation = async (accessToken, stationId) => {
+  const url = `${BASE_URL}/stations/${stationId}`;
+  const option = {
+    method: 'DELETE',
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      'Content-Type': 'application/json',
+    },
+  };
+
+  try {
+    await request(url, option);
+  } catch (error) {
+    throw new Error(error);
+  }
+};
+
+export const stationManageAPI = {
+  getStations,
+  addStation,
+  modifyStation,
+  removeStation,
+};
