+++ src/js/subway/utils/index.js
@@ -1,4 +1,6 @@
 export { isValidEmail, isValidName, isValidPassword, findInValidInput } from './validate';
-export { getUserName } from './getUserName';
 export { routeTo, getRedirectedPath } from './route';
 export { showModal, hideModal } from './modal';
+export { userJoinAPI } from './api/userJoin';
+export { userAuthAPI } from './api/userAuth';
+export { stationManageAPI } from './api/stationManage';
